--- a/AK/NumericLimits.h
+++ b/AK/NumericLimits.h
@@ -19,6 +19,7 @@ struct NumericLimits<bool> {
     static constexpr bool min() { return false; }
     static constexpr bool max() { return true; }
     static constexpr bool is_signed() { return false; }
+    static constexpr size_t digits() { return 1; }
 };
 
 template<>
@@ -26,6 +27,7 @@ struct NumericLimits<signed char> {
     static constexpr signed char min() { return -__SCHAR_MAX__ - 1; }
     static constexpr signed char max() { return __SCHAR_MAX__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __CHAR_BIT__ - 1; }
 };
 
 template<>
@@ -33,6 +35,7 @@ struct NumericLimits<char> {
     static constexpr char min() { return -__SCHAR_MAX__ - 1; }
     static constexpr char max() { return __SCHAR_MAX__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __CHAR_BIT__ - 1; }
 };
 
 template<>
@@ -40,6 +43,7 @@ struct NumericLimits<short> {
     static constexpr short min() { return -__SHRT_MAX__ - 1; }
     static constexpr short max() { return __SHRT_MAX__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(short) - 1; }
 };
 
 template<>
@@ -47,6 +51,7 @@ struct NumericLimits<int> {
     static constexpr int min() { return -__INT_MAX__ - 1; }
     static constexpr int max() { return __INT_MAX__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(int) - 1; }
 };
 
 template<>
@@ -54,6 +59,7 @@ struct NumericLimits<long> {
     static constexpr long min() { return -__LONG_MAX__ - 1; }
     static constexpr long max() { return __LONG_MAX__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(long) - 1; }
 };
 
 template<>
@@ -61,6 +67,7 @@ struct NumericLimits<long long> {
     static constexpr long long min() { return -__LONG_LONG_MAX__ - 1; }
     static constexpr long long max() { return __LONG_LONG_MAX__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(long long) - 1; }
 };
 
 template<>
@@ -68,6 +75,7 @@ struct NumericLimits<unsigned char> {
     static constexpr unsigned char min() { return 0; }
     static constexpr unsigned char max() { return __SCHAR_MAX__ * 2u + 1; }
     static constexpr bool is_signed() { return false; }
+    static constexpr size_t digits() { return __CHAR_BIT__; }
 };
 
 template<>
@@ -75,6 +83,7 @@ struct NumericLimits<unsigned short> {
     static constexpr unsigned short min() { return 0; }
     static constexpr unsigned short max() { return __SHRT_MAX__ * 2u + 1; }
     static constexpr bool is_signed() { return false; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(short); }
 };
 
 template<>
@@ -82,6 +91,7 @@ struct NumericLimits<unsigned> {
     static constexpr unsigned min() { return 0; }
     static constexpr unsigned max() { return __INT_MAX__ * 2u + 1; }
     static constexpr bool is_signed() { return false; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(int); }
 };
 
 template<>
@@ -89,6 +99,7 @@ struct NumericLimits<unsigned long> {
     static constexpr unsigned long min() { return 0; }
     static constexpr unsigned long max() { return __LONG_MAX__ * 2ul + 1; }
     static constexpr bool is_signed() { return false; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(long); }
 };
 
 template<>
@@ -96,6 +107,7 @@ struct NumericLimits<unsigned long long> {
     static constexpr unsigned long long min() { return 0; }
     static constexpr unsigned long long max() { return __LONG_LONG_MAX__ * 2ull + 1; }
     static constexpr bool is_signed() { return false; }
+    static constexpr size_t digits() { return __CHAR_BIT__ * sizeof(long long); }
 };
 
 #ifndef KERNEL
@@ -106,6 +118,7 @@ struct NumericLimits<float> {
     static constexpr float max() { return __FLT_MAX__; }
     static constexpr float epsilon() { return __FLT_EPSILON__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __FLT_MANT_DIG__; }
 };
 
 template<>
@@ -115,6 +128,7 @@ struct NumericLimits<double> {
     static constexpr double max() { return __DBL_MAX__; }
     static constexpr double epsilon() { return __DBL_EPSILON__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __DBL_MANT_DIG__; }
 };
 
 template<>
@@ -124,6 +138,7 @@ struct NumericLimits<long double> {
     static constexpr long double max() { return __LDBL_MAX__; }
     static constexpr long double epsilon() { return __LDBL_EPSILON__; }
     static constexpr bool is_signed() { return true; }
+    static constexpr size_t digits() { return __LDBL_MANT_DIG__; }
 };
 #endif