This contains two performance improvements for Deflate decompression:
* Increase `LittleEndianInputBitStream` bit buffer size to a u64 and replace loops with bitwise operations
* Use a prefix table to decode Huffman codes up to 8 bits in length

Overall, this decreases decoding of the [enwik8](https://www.mattmahoney.net/dc/enwik8.zip) file (100MB when uncompressed) from:
* 11.1s to 1.6s on my Linux machine
* 242s to 5.8s on Serenity (cold)
* 242s to 2.3s on Serenity (warm)